ADD: pull out handle login into own file
