+++ controller/userController.js
@@ -15,28 +15,28 @@ const bcrypt =  require('bcrypt')
 const jwt =  require('jsonwebtoken')
 require('dotenv').config()
 
-const handleRegistration = (req, res) => {
-    const { username, password } =  req.body
-    // check if not blank
-    if(!username || !password) res.status(400).json({'message': 'Username & Password are required'})
-    //check of existsnce
-    const duplicate = UsersDB.users.find((person) => person.username === username)
-    if(duplicate) res.sendStatus(409) // conflict
-    // encrypt and store
-    const saltRounds = 10
-    bcrypt.genSalt(saltRounds, function(err, salt) {
-        bcrypt.hash(password, salt, function(err, hash) {
-            //catch any errors
-            if(err) {
-                res.sendStatus(500)
-            }else{
-                const newUser = {id: uuidv4(), username, password: hash}
-                UsersDB.setUsers([...UsersDB.users, newUser])
-                res.status(201).json({'message': `${newUser.username} created`})
-            }
-        });
-    });
-}
+// const handleRegistration = (req, res) => {
+//     const { username, password } =  req.body
+//     // check if not blank
+//     if(!username || !password) res.status(400).json({'message': 'Username & Password are required'})
+//     //check of existsnce
+//     const duplicate = UsersDB.users.find((person) => person.username === username)
+//     if(duplicate) res.sendStatus(409) // conflict
+//     // encrypt and store
+//     const saltRounds = 10
+//     bcrypt.genSalt(saltRounds, function(err, salt) {
+//         bcrypt.hash(password, salt, function(err, hash) {
+//             //catch any errors
+//             if(err) {
+//                 res.sendStatus(500)
+//             }else{
+//                 const newUser = {id: uuidv4(), username, password: hash}
+//                 UsersDB.setUsers([...UsersDB.users, newUser])
+//                 res.status(201).json({'message': `${newUser.username} created`})
+//             }
+//         });
+//     });
+// }
 
 const handleLogin = (req, res) => {
     const { username, password } = req.body
@@ -77,12 +77,12 @@ const genAccessToken = (req, res) => {
     const refreshToken =  cookies.jwt
     const foundUser =  UsersDB.users.find((user) => user.refreshToken === refreshToken)
     if(!foundUser) return res.sendStatus(403)
-
+    // console.log(foundUser)
     jwt.verify(
         {"username": foundUser.username},
         process.env.REFRESH_TOKEN_SECRET,
         (err, decoded) => {
-            if(err || foundUser.username !== decoded.username) res.sendStatus(403)
+            if(err || foundUser.username !== decoded.username) return res.sendStatus(403)
             const accessToken =  jwt.sign(
                 {"username": decoded.username},
                 process.env.ACCESS_TOKEN_SECRET,
@@ -116,7 +116,7 @@ const handleLogout = (req, res) => {
 }
 
 module.exports = { 
-    handleRegistration,
+    // handleRegistration,
     handleLogin,
     genAccessToken,
     handleLogout
